Add files via upload
Here's an example of a Python code for a fancy font generator tool using the pyfiglet library:
In this code, we're using the pyfiglet library to generate fancy ASCII art text based on user input. Here's how it works:

Install the pyfiglet library by running pip install pyfiglet in your terminal.

The generate_fancy_text() function takes a text parameter as input and uses pyfiglet.figlet_format() to generate fancy ASCII art text. It returns the generated text.

In the main section of the code, we prompt the user to enter their desired text using input(). This text is then passed to the generate_fancy_text() function to generate the fancy text.

Finally, the generated fancy text is printed to the console using print().
You can check the Fancy Fonts, generated by this tool on https://www.socialtricks.net/